+++ back-end/products/models/__init__.py
@@ -1,0 +1,5 @@
+from .bank import Bank, BankOption
+from .DepositProduct import DepositProduct, DepositProductOption
+from .SavingProduct import SavingProduct, SavingProductOption
+from .PensionProduct import PensionProduct, PensionProductOption
+from .RentLoanProduct import RentLoanProduct, RentLoanProductOption
